Rename store file and type
