+++ src/frontend/screens/Game/GamePage/components/SettingsButton.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { GameInfo } from 'common/types'
 import { SvgButton } from 'frontend/components/UI'
 import SettingsIcoAlt from 'frontend/assets/settings_icon_alt.svg?react'
-import { openGameSettingsModal } from 'frontend/state/SettingsModal'
+import { openGameSettingsModal } from 'frontend/state/GameSettingsModal'
 
 interface Props {
   gameInfo: GameInfo
