+++ src/frontend/screens/Game/GamePage/index.tsx
@@ -67,7 +67,7 @@ import { hasKnownFixes } from 'frontend/hooks/hasKnownFixes'
 import { openInstallGameModal } from 'frontend/state/InstallGameModal'
 import useSettingsContext from 'frontend/hooks/useSettingsContext'
 import SettingsContext from 'frontend/screens/Settings/SettingsContext'
-import { useGameSettingsModal } from 'frontend/state/SettingsModal'
+import { useGameSettingsModal } from 'frontend/state/GameSettingsModal'
 
 export default React.memo(function GamePage(): JSX.Element | null {
   const { appName, runner } = useParams() as { appName: string; runner: Runner }
